#121 simplified artwork scan to be in the same process of audio scan
