+++ src/cmd/socket/nodepb/messages_pb.js
@@ -668,8 +668,7 @@ proto.pb.ArtworkExtractInput.prototype.toObject = function(opt_includeInstance) 
 proto.pb.ArtworkExtractInput.toObject = function(includeInstance, msg) {
   var f, obj = {
     media: (f = msg.getMedia()) && media_pb.Media.toObject(includeInstance, f),
-    scanDate: (f = msg.getScanDate()) && google_protobuf_timestamp_pb.Timestamp.toObject(includeInstance, f),
-    root: jspb.Message.getFieldWithDefault(msg, 3, "")
+    scanDate: (f = msg.getScanDate()) && google_protobuf_timestamp_pb.Timestamp.toObject(includeInstance, f)
   };
 
   if (includeInstance) {
@@ -716,10 +715,6 @@ proto.pb.ArtworkExtractInput.deserializeBinaryFromReader = function(msg, reader)
       reader.readMessage(value,google_protobuf_timestamp_pb.Timestamp.deserializeBinaryFromReader);
       msg.setScanDate(value);
       break;
-    case 3:
-      var value = /** @type {string} */ (reader.readString());
-      msg.setRoot(value);
-      break;
     default:
       reader.skipField();
       break;
@@ -763,13 +758,6 @@ proto.pb.ArtworkExtractInput.serializeBinaryToWriter = function(message, writer)
       2,
       f,
       google_protobuf_timestamp_pb.Timestamp.serializeBinaryToWriter
-    );
-  }
-  f = message.getRoot();
-  if (f.length > 0) {
-    writer.writeString(
-      3,
-      f
     );
   }
 };
@@ -846,24 +834,6 @@ proto.pb.ArtworkExtractInput.prototype.clearScanDate = function() {
  */
 proto.pb.ArtworkExtractInput.prototype.hasScanDate = function() {
   return jspb.Message.getField(this, 2) != null;
-};
-
-
-/**
- * optional string root = 3;
- * @return {string}
- */
-proto.pb.ArtworkExtractInput.prototype.getRoot = function() {
-  return /** @type {string} */ (jspb.Message.getFieldWithDefault(this, 3, ""));
-};
-
-
-/**
- * @param {string} value
- * @return {!proto.pb.ArtworkExtractInput} returns this
- */
-proto.pb.ArtworkExtractInput.prototype.setRoot = function(value) {
-  return jspb.Message.setProto3StringField(this, 3, value);
 };
 
 
@@ -5926,7 +5896,6 @@ proto.pb.Message = {
   MESSAGE_MEDIA_SPECTRUM_GENERATE: 18,
   MESSAGE_MEDIA_DB_SELECT: 19,
   MESSAGE_MEDIA_EXTRACT_ARTWORK_FROM_FILE: 20,
-  MESSAGE_MEDIA_EXTRACT_ARTWORK_FROM_DIRECTORIES: 21,
   MESSAGE_IPC: 100,
   MESSAGE_MPV_EOF_REACHED: 202,
   MESSAGE_MPV_PERCENT_POS: 203,
