+++ ui-player/src/api/index.js
@@ -20,7 +20,6 @@ const audioScan = data => axios.post(`/audio/scan`, data);
 const albumViewList = data => axios.get(`/media/albums`, {
     params: data,
 })
-const artworkScan = data => axios.post(`/artwork/scan`, data);
 const decodeURL = search => querystring.decode(search.replace('?',''));
 const directoryRead = async data => {
     const response = await axios.post(`/file-browser/directory`, data);
@@ -47,7 +46,6 @@ const spectrumImage = (id) => axios.get(`/media/${id}/spectrum`);
 export default {
     audioScan,
     albumViewList,
-    artworkScan,
     decodeURL,
     directoryRead,
     distinctListGet,
